tasks

+++ src/main/java/com/stackroute/service/TrackServiceImpl.java
@@ -1,6 +1,8 @@
 package com.stackroute.service;
 
 import com.stackroute.domain.Track;
+import com.stackroute.exceptions.TrackAlreadyExistsException;
+import com.stackroute.exceptions.TrackNotFoundException;
 import com.stackroute.repository.TrackRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,8 +20,17 @@ public class TrackServiceImpl implements TrackService {
     }
 
     @Override
-    public Track saveTrack(Track track) {
-        return trackRepository.save(track);
+    public Track saveTrack(Track track) throws TrackAlreadyExistsException {
+        if(trackRepository.existsById(track.getTrackId()))
+        {
+            throw new TrackAlreadyExistsException("Track already exists");
+        }
+        Track savedTrack = trackRepository.save(track);
+        if(savedTrack==null)
+        {
+            throw new TrackAlreadyExistsException("Track already exists");
+        }
+        return savedTrack;
     }
 
     @Override
@@ -34,12 +45,21 @@ public class TrackServiceImpl implements TrackService {
     }
 
     @Override
-    public void removeTrack(int trackId) {
+    public void removeTrack(int trackId) throws TrackNotFoundException {
+        if(!trackRepository.existsById(trackId))
+        {
+            throw new TrackNotFoundException("Track not found");
+        }
         trackRepository.deleteById(trackId);
     }
 
     @Override
-    public Track trackByName(String trackName) {
+    public Track trackByName(String trackName) throws TrackNotFoundException {
+        if(trackRepository.findByTitle(trackName)==null)
+        {
+            throw new TrackNotFoundException("Track not found");
+        }
         return trackRepository.findByTitle(trackName);
+
     }
 }
